feat: D-Day Long 타입으로 변경

+++ src/main/java/com/on/server/domain/diary/dto/DiaryDto.java
@@ -13,5 +13,5 @@ public class DiaryDto {
 
     private String content;
 
-    private int writtenDday;
+    private Long writtenDday;
 }
